model.py
import numpy as np
import pandas as pd
from tensorflow.keras.models import load_model
from PIL import Image
import io

# Load your trained model
model = load_model('my_cnn_model.keras')

def process_image(file):
    # User-defined image processing logic
    image = Image.open(file.stream)
    # Example processing: resize and convert to array
    image = image.resize((160, 40))  # Adjust to your model's input size
    image_array = np.array(image) / 255.0  # Normalize if needed
    # Convert to a DataFrame or any other required structure
    df = pd.DataFrame(image_array.flatten()).T  # Adjust as necessary for your model
    return df

def predict(df):
    # Make predictions
    prediction = model.predict(df)
    return np.argmax(prediction, axis=1)  # Return the predicted class
