requirements.txt
pandas
matplotlib
seaborn
scikit-learn
tensorflow
Flask
pillow
opencv-python